Update draft-ietf-ccwg-rfc5033bis.md
I forgot to merge the last set of minor changes to help improve flow and expalin why we had these sections. 
And one undo of my chnage: I suggest that for a proposal entering the IETF, proposers need to be aware of security, not to actually answer how security is impacted -  to me.that's a part of the IETF process.