Minor update to fix NiTs from Shepherd Review
A new PR will addreess these issues, when rev'ed as an ID.

1. -- Obsolete informational reference (is this intentional?): RFC 2988
     (Obsoleted by RFC 6298)

mea culpa for not spotting - we ought to replace by the later RFC.

2.  -- Obsolete informational reference (is this intentional?): RFC 8312
     (Obsoleted by RFC 9438)

That was intentional. The text of the PR has been updated to say this.

3. We can add a note that tunnels, and similar encaps need to consider nested congestion control interactions:

The design of tunnels and similar encapsulations might need to consider nested congestion control interactions. For example, when ECN is used by an IP and lower layer technology {{ECN-encaps}}.